+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Migration**) Sort destination source selection by same order as browse source page
 - (**Migration**) Sort entries in search/execution page by title
 - (**Migration**) Prevent a manual selected match from getting automatically overwritten by a new-found match
+- (**Migration**) Abort active search when selecting a match through the manual search
 
 ### Fixed